Proof of concept for variable based subsites
Implemented sytem to allow settings files that prevent the need for template modification on ward subsites, meaning merging future changes across the network will be simple.
Essentially there are now only 2 files that need editing for a ward subsite:
_includes/ward-settings.njk
_includes/ward-resources.njk

If others see this as the way forward I will implment a more substantial system to allow further ward modifications to be managed in this way.